Shift times profiles fixes,
text editor wrapping and drawing fixes,
bidi conversion fixes.

+++ Kaiplayer/Provider.h
@@ -63,7 +63,7 @@ public:
 	bool AudioNotInitialized() {
 		return audioNotInitialized;
 	}
-	int GetAudioProgress() {
+	float GetAudioProgress() {
 		return m_audioProgress;
 	}
 protected:
